Working example, but completed to 85 - 90 %
Multiple threads with factory function for the sensor array. Multiple synchronicity.
Sensor readings are based on a Gaussian integration processes, so the output looks like a signal.
Sensor messages are delivered in JSON format with current POSIX timestamp.
Missing the pydantic verifiers for the repositories.
Repositories are not working yet. finer points on db and file creation.